Add files via upload
zabranie mozliwosci logowania jesli żaden użytkownik nie istnieje oraz potwierdzanie wyborow

+++ main.cpp
@@ -2,37 +2,38 @@
 #include "database.h"
 #include <conio.h>
 #include <stdlib.h>
-void menu() {
 
+
+
+
+bool confirmation() {
+	std::cout << "------------------------------------------------------------\n";
+	cout << "Czy chcesz kontynuowac(y)?\n";
+	std::cout << "------------------------------------------------------------\n";
+	char choice;
+	choice = _getch();
+	switch (choice)
+	{
+	case 'y':
+		return true;
+		break;
+	case 'Y':
+		return true;
+		break;
+	default:
+		return false;
+		break;
+	}
 }
+
+
 int main()
 {
+	
 	char choice{};
 	int attempts = 1;
 	Database db;
-	std::cout << "Menedzer hasel\n\n1.Zaloguj sie\n2.Zarejestruj sie\n";
-	choice = _getch();
-
-	
-	switch (choice)
-	{
-	case '1':
-		do {
-			if (!db.logged && attempts > 1) std::cout << "\nNieprawidlowe dane\n";
-			db.loginToAccount();
-			attempts++;
-		} while (db.logged != true);
-		break;
-	case '2':
-		do {
-			db.createAccount();
-			attempts++;
-		} while (db.registered != true);
-		break;
-	default:
-		std::cout << "Nieprawidlowy znak\n";
-		break;
-	}
+	db.menuToDatabase();
 
 	if (db.logged  || db.registered) 
 	{
@@ -41,36 +42,48 @@ int main()
 		{
 			std::cout << "Menedzer hasel\n\n1.Dodaj haslo\n2.Usun haslo\n3.Zmien Haslo\n4.Wyswietl wszystkie hasla\n5.Wyswietl haslo po id lub nazwie\n6.Zamknij program\n";
 			choice = _getch();
-			std::cout << "------------------------------------------------------------" << std::endl;
-			switch (choice)
-			{
-			case '1':
-				db.insertData();
-				break;
-			case '2':
-				db.deleteData();
-				break;
-			case '3':
-				db.changePassword();
-				break;
-			case '4':
-				db.selectData();
-				break;
-			case '5':
-				db.selectOne();
-				break;
-			case '6':
-				exit(1);
-
-			default:
-				std::cout << "Nieprawidlowy znak\n";
-				break;
-			}
-			std::cout << "------------------------------------------------------------" << std::endl;
-			std::cout << "Nacisnij Enter aby kontynuowac" << std::endl;
-			_getch();
-			system("cls");
 			
+				switch (choice)
+				{
+				case '1':
+					if (confirmation()) {
+						db.insertData();
+					}
+					break;
+				case '2':
+					if (confirmation()) {
+						db.deleteData();
+					}
+					break;
+				case '3':
+					if (confirmation()) {
+						db.changePassword();
+					}
+					break;
+				case '4':
+					if (confirmation()) {
+						db.selectData();
+					}
+					break;
+				case '5':
+					if (confirmation()) {
+						db.selectOne();
+					}
+					break;
+				case '6':if (confirmation()) {
+					exit(1);
+				}
+
+				default:
+					std::cout << "Nieprawidlowy znak\n";
+					break;
+				}
+				std::cout << "------------------------------------------------------------" << std::endl;
+				std::cout << "Nacisnij Enter aby kontynuowac" << std::endl;
+				_getch();
+				system("cls");
+			
+
 		}
 	}
 	system("pause");
